% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acct_achievement.R
\name{acct_achievement}
\alias{acct_achievement}
\title{Accountability Achievement Calculations}
\usage{
acct_achievement(sl_path, grade_pools_path, cte_alt_adult_path,
  school_names_path, success_amo_path, act_substitution_path,
  school_assessment_path, a_cut, b_cut, c_cut, d_cut)
}
\arguments{
\item{sl_path}{Path to the Student Level file}

\item{grade_pools_path}{Path to the Grade Pools file}

\item{school_names_path}{Path to the file with School/System names}

\item{success_amo_path}{Path to the AMO file for the Achievement Indicator}

\item{act_substitution_path}{Path to the ACT Substitution file}

\item{school_assessment_path}{Path to the school level Assessment file}

\item{a_cut}{Cut score for 'A' in Absolute Pathway}

\item{b_cut}{Cut score for 'B' in Absolute Pathway}

\item{c_cut}{Cut score for 'C' in Absolute Pathway}

\item{d_cut}{Cut score for 'D' in Absolute Pathway}
}
\description{
This function creates Accountability dataframe for the Achievement Indicator
}
\examples{
acct_achievement("N://ORP_accountability/projects/2019_student_level_file/2019_student_level_file.csv",
"N:/ORP_accountability/projects/2019_school_accountability/grade_pools_designation_immune.csv",
"N:/ORP_accountability/data/2019_final_accountability_files/names.csv",
"N:/ORP_accountability/projects/2019_amo/success_rate_targets_school.csv",
"N:/ORP_accountability/data/2019_final_accountability_files/act_substitution_school.csv",
"N:/ORP_accountability/data/2019_final_accountability_files/school_assessment_file.csv",
45, 35, 27.5, 20)
}
\keyword{accountability}
\keyword{achievement,}
\keyword{indicator,}
