% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acct_elpa.R
\name{acct_elpa}
\alias{acct_elpa}
\title{Accountability ELPA Calculations}
\usage{
acct_elpa(school_elpa_path, grade_pools_path, school_names_path,
  a_cut = 60, b_cut = 50, c_cut = 40, d_cut = 25,
  min_n_count = 10)
}
\arguments{
\item{school_elpa_path}{Path to the School Level ELPA file}

\item{grade_pools_path}{Path to the Grade Pools file}

\item{school_names_path}{Path to the file with School/System names}

\item{a_cut}{Cut score for 'A' in Absolute Pathway}

\item{b_cut}{Cut score for 'B' in Absolute Pathway}

\item{c_cut}{Cut score for 'C' in Absolute Pathway}

\item{d_cut}{Cut score for 'D' in Absolute Pathway}

\item{min_n_count}{Minimum N Count needed to receive score}
}
\description{
This function creates Accountability dataframe for the ELPA Indicator
}
\examples{
acct_elpa("N:/ORP_accountability/data/2019_ELPA/wida_growth_standard_school.csv",
"N:/ORP_accountability/projects/2019_school_accountability/grade_pools_designation_immune.csv",
"N:/ORP_accountability/data/2019_final_accountability_files/names.csv",
60, 50, 40, 25, min_n_count= 10)
}
\keyword{accountability}
\keyword{el,}
\keyword{elpa,}
\keyword{english}
\keyword{indicator,}
\keyword{learner,}
