% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acct_grad_rate.R
\name{acct_grad_rate}
\alias{acct_grad_rate}
\title{Accountability Achievement Calculations}
\usage{
acct_grad_rate(
  school_grad_rate_path,
  grade_pools_path,
  school_names_path,
  grad_amo_path,
  a_cut = 95,
  b_cut = 90,
  c_cut = 80,
  d_cut = 67,
  min_n_count = 30
)
}
\arguments{
\item{school_grad_rate_path}{Path to the School Level Graduation Rate file}

\item{grade_pools_path}{Path to the Grade Pools file}

\item{school_names_path}{Path to the file with School/System names}

\item{grad_amo_path}{Path to the AMO file for the Graduation Rate Indicator}

\item{a_cut}{Cut score for 'A' in Absolute Pathway}

\item{b_cut}{Cut score for 'B' in Absolute Pathway}

\item{c_cut}{Cut score for 'C' in Absolute Pathway}

\item{d_cut}{Cut score for 'D' in Absolute Pathway}

\item{min_n_count}{Minimum N Count needed to receive score}
}
\description{
This function creates Accountability dataframe for the Graduation Rate Indicator
}
\examples{
\dontrun{
acct_grad_rate("N:/ORP_accountability/data/2018_graduation_rate/school_grad_rate.csv",
"N:/ORP_accountability/projects/2019_school_accountability/grade_pools_designation_immune.csv",
"N:/ORP_accountability/data/2019_final_accountability_files/names.csv",
"N:/ORP_accountability/projects/2019_amo/grad_school.csv",
95, 90, 80, 67, min_n_count= 30)
}

}
\keyword{accountability}
\keyword{graduation}
\keyword{indicator,}
\keyword{rate,}
