% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acct_tvaas_growth.R
\name{acct_tvaas_growth}
\alias{acct_tvaas_growth}
\title{Accountability TVAAS (Growth) Calculations}
\usage{
acct_tvaas_growth(tvaas_path_excel, grade_pools_path, school_names_path)
}
\arguments{
\item{tvaas_path_excel}{Path to the TVAAS file from SAS in Excel format}

\item{grade_pools_path}{Path to the Grade Pools file}

\item{school_names_path}{Path to the file with School/System names}
}
\description{
This function creates Accountability dataframe for the Growth Indicator
}
\examples{
acct_tvaas_growth("N:/ORP_accountability/data/2019_tvaas/2019-School-Level-Accountability-Results-EOC-TCAP.xlsx",
"N:/ORP_accountability/projects/2019_school_accountability/grade_pools_designation_immune.csv",
"N:/ORP_accountability/data/2019_final_accountability_files/names.csv")
}
\keyword{accountability}
\keyword{growth,}
\keyword{indicator,}
