% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_student_level.R
\name{agg_student_level}
\alias{agg_student_level}
\title{Aggregate Student Level to School/District/State}
\usage{
agg_student_level(df, sum_cols, mean_cols, ...)
}
\arguments{
\item{df}{Dataframe. Used first to be friendly to the \%>\% operator.}

\item{sum_cols}{A character vector of the columns to be summed.}

\item{mean_cols}{A character vector of the columns to find the mean. Use empty vector if no desired mean columns.}

\item{...}{The columns to aggregate at (e.g. system, system_name, school, school_name)}
}
\description{
This function takes a student level dataframe and aggregates it to the school/district/state level.
Provides sums and means for desired columns.
}
\examples{
agg_student_level(student_level_df, c('enrolled', 'tested', 'valid tests'), c(), system, system_name, subgroup)
agg_student_level(student_level_df, c('enrolled', 'tested', 'valid tests'), c(), system, school, subgroup)
}
\keyword{aggregate,}
\keyword{collapse}
\keyword{group,}
