% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pcts.R
\name{calc_pcts}
\alias{calc_pcts}
\title{Calculates Percentages}
\usage{
calc_pcts(df, numerator_cols, denom_col)
}
\arguments{
\item{df}{Dataframe. Used first to be friendly to the \%>\% operator.}

\item{numerator_cols}{A character vector of columns to be used in the numerator.}

\item{denom_col}{A single string of the column name to be used in the denominator.}
}
\description{
This function takes a dataframe and a list of columns to find a percentage for, along with a single column to use
in the denominator. It returns a dataframe
}
\examples{
\dontrun{
calc_pcts(system_level, c('on_track', 'mastered'), 'valid_tests')
}

}
\keyword{calculate}
\keyword{percent,}
\keyword{percentage,}
