% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_interval.R
\name{confidence_interval}
\alias{confidence_interval}
\title{A Confidence Interval Function}
\usage{
confidence_interval(df, metric_column = "metric",
  denom_column = "n_count", interval = 95, bound = "upper")
}
\arguments{
\item{df}{Dataframe. Used first to be friendly to the \%>\% operator.}

\item{metric_column}{The column name desired for confidence interval. Defaults to metric}

\item{denom_column}{Column name of the denominator used. Defaults to n_count}

\item{interval}{The percent confidence for the interval. Defaults to 95.}

\item{bound}{The side of the confidence interval. Defaults to upper. Choices are upper or lower.}

\item{out_column}{The name for the resulting column. Defaults to ci_bound}
}
\description{
This function creates a new column with the desired confidence interval bound. Names the new column ci_bound.
}
\examples{
confidence_interval()
}
\keyword{ci}
\keyword{confidence}
\keyword{confidence,}
\keyword{interval,}
