% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tdoe.R
\name{save_with_logo}
\alias{save_with_logo}
\title{Save plot with TDOE Logo
Default is the size of image determined to work best with PowerPoint slides}
\usage{
save_with_logo(file_path, plot_object = ggplot2::last_plot(),
  fig_height = 4.95, fig_width = 9.17, fig_unit = "in",
  fig_res = 1200)
}
\arguments{
\item{file_path}{Path to save the image, ending in .png}

\item{plot_object}{ggplot object to save}

\item{fig_height}{Height of image. Default of 4.95}

\item{fig_width}{Width of image. Default of 9.17}

\item{fig_unit}{Default of 'in' for inches. Can use 'px' if units are pixels}

\item{fig_res}{Resolution of image. Default of 1200.}
}
\description{
Save plot with TDOE Logo
Default is the size of image determined to work best with PowerPoint slides
}
