% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tdoe.R
\name{tdoe_pal}
\alias{tdoe_pal}
\title{TDOE color palette}
\usage{
tdoe_pal()
}
\description{
The standard three-color TDOE palette for line plots comprises
red, navy, and gray
}
\seealso{
Other colour tdoe: 
\code{\link{scale_colour_tdoe}()}
}
\concept{colour tdoe}
