% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tdoe.R
\name{theme_tdoe}
\alias{theme_tdoe}
\title{Theme for TDOE plots}
\usage{
theme_tdoe(base_size = 9, base_family = "sans")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Theme for TDOE plots
}
\concept{themes tdoe}
